requirements.txt
fastapi
uvicorn
requests
pillow
numpy
jinja2
python-multipart